% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDMR2.R
\name{agregation}
\alias{agregation}
\title{\code{agregation} aggregates all the experts preferences following and Yager OWA quantifier}
\usage{
agregation(preferences, M, N, quantifier)
}
\arguments{
\item{preferences}{amtrix with the experts preferences}

\item{M}{Number of experts}

\item{N}{Number of aternatives}

\item{quantifier}{Type of quantifier could be 'leasthalf', 'most', 'asmany'}
}
\value{
A NxN matrix with the preferences of all the experts aggregated using the given quantifier
}
\description{
\code{agregation} aggregates all the experts preferences following and Yager OWA quantifier
}

