% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{Title complete
\code{complete} Completes the missing preferences values in a matrix of preferences following the additive consistency property}
\usage{
complete(preferences)
}
\arguments{
\item{preferences}{a matrix where all the incomplete experts preferences are included, that is, if there are n alternatives and h experts it receives a matrix of
h*n rows and n cols}
}
\value{
the complete matrix of preferences
}
\description{
Title complete
\code{complete} Completes the missing preferences values in a matrix of preferences following the additive consistency property
}

