% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDMR2.R
\name{distanceCalculation}
\alias{distanceCalculation}
\title{Calculates the distance between two elements}
\usage{
distanceCalculation(a, b, type)
}
\arguments{
\item{a}{Firt element}

\item{b}{Second element}

\item{type}{Type of distance, could be 'manhattan', 'euclidean', 'cosine', 'dice', 'jacard'}
}
\value{
d the distance between the two given elements
}
\description{
Calculates the distance between two elements
}

