% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{estimate}
\alias{estimate}
\title{Estimates the missing value using the additive consistency properties}
\usage{
estimate(i, k, preferences, type)
}
\arguments{
\item{i}{number of row where the preference is located}

\item{k}{number of column where the preference is located}

\item{preferences}{matrix of preferences}

\item{type}{type of estimation}
}
\value{
the estimated value
}
\description{
Estimates the missing value using the additive consistency properties
}

