% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDMR2.R
\name{exploitation}
\alias{exploitation}
\title{Exploitation}
\usage{
exploitation(agregated, N, dominance, quantifier)
}
\arguments{
\item{agregated}{NxN matrix of preference where all the experts preferences has been included}

\item{N}{Number of expert}

\item{dominance}{Type of choice degree "GDD" or "GNDD"}

\item{quantifier}{Type of quantifier}
}
\description{
Exploitation
}

