% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferencesModule.R
\name{generateIncompletePref}
\alias{generateIncompletePref}
\title{generateIncompletePref Generates a matrix of incomplete preferences ramdomly colocated}
\usage{
generateIncompletePref(n, consistent)
}
\arguments{
\item{n}{number of alternatives}

\item{consistent}{boolean paramater that indicates if the maxtrix of preference is consistent or not}
}
\value{
pref a matrix of preferences
}
\description{
generateIncompletePref Generates a matrix of incomplete preferences ramdomly colocated
}

