% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferencesModule.R
\name{generatePref}
\alias{generatePref}
\title{generatePref Generates a matrix of preferences}
\usage{
generatePref(n, consistent)
}
\arguments{
\item{n}{number of alternatives}

\item{consistent}{boolean paramater that indicates if the maxtrix of preference is consistent or not}
}
\value{
pref a matrix of preferences
}
\description{
generatePref Generates a matrix of preferences
}

