% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDMR2.R
\name{owa}
\alias{owa}
\title{owa Aggregates the elementes in the vector A following an Ordering Weighting Averaging operator, OWA, with the weights provided as a parameter}
\usage{
owa(A, w)
}
\arguments{
\item{A}{number of alternatives}

\item{w}{Weights for the aggregation}
}
\value{
h Aggregated value
}
\description{
owa Aggregates the elementes in the vector A following an Ordering Weighting Averaging operator, OWA, with the weights provided as a parameter
}

