% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDMR2.R
\name{weight}
\alias{weight}
\title{Computes the weights in the aggregation when a Yager Ordering weighting averaging operator (OWA) is used}
\usage{
weight(r, type)
}
\arguments{
\item{type}{Type of quantifier to be used in the aggregation, could be 'leasthalf', 'most' and 'asmany'.}
}
\value{
weight for the BUM function
}
\description{
Computes the weights in the aggregation when a Yager Ordering weighting averaging operator (OWA) is used
}

